Add new Unused Indexes query from Cybertec
Purpose: pghero is currently reporting that we have no unused indexes.
However when I use this Cybertec query that attempts to report on unused
indexes, we have many. So it is possible the current pghero unused
indexes query is missing unused indexes and that seems like a valuable
thing to add in.

Credit: https://www.cybertec-postgresql.com/en/get-rid-of-your-unused-indexes/

It has at least one more WHERE condition compared with the current
pghero unused indexes query in that it looks at whether the index
enforces a constraint or not.

Cybertec version also appears to check "no index column is an
expression"

I've been checking the idx_scan = 0 as well, although I adapted it
slightly to preserve pghero passing in a `max_scans` which has a default
of 50

Both versions have in common: excluding UNIQUE indexes