Scheduled Jobs
Relocate the methods away from query stats where they are not related

Add more "blank slate" handling which could include the extension not
being available, the extension being available but not enabled, or the
extension being enabled. In addition to that another state would be that
the extension is enabled but there are no scheduled jobs.

Add routes to enable the extension when it is available but not enabled,
from the web UI.

Update some of the Guide page.